# TicketManagementSystem/backend/TicketManagementSystem.API/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Management System API
  description: |
    API RESTful para el Sistema de Gestión de Tickets.
    
    ## Autenticación
    Esta API utiliza **JWT Bearer Token** para autenticación. Para acceder a los endpoints protegidos:
    1. Obtener un token mediante `/api/auth/login`
    2. Incluir el token en el header `Authorization: Bearer {token}`
    
    ## Rate Limiting
    - General: 200 requests/minuto
    - Login: 20 requests/minuto
    - Register: 10 requests/minuto
    
    ## Roles
    - **Admin**: Acceso completo al sistema
    - **Agent**: Gestión de tickets asignados
    - **User**: Crear y ver tickets propios
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@ticketmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
  - url: https://localhost:5001/api
    description: Servidor de desarrollo (HTTPS)
  - url: https://api.ticketmanagement.com/api
    description: Servidor de producción
...
security:
  - BearerAuth: []
paths:
...
  /comments:
    post:
      tags:
        - Comments
      summary: Crear comentario
      description: Agregar un comentario a un ticket
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
            example:
              ticketId: 1
              content: "He revisado el error y parece ser un problema de configuración"
      responses:
        '201':
          description: Comentario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Ticket no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /comments/{id}:
    put:
      tags:
        - Comments
      summary: Actualizar comentario
      description: Actualizar un comentario existente (solo autor)
      operationId: updateComment
      parameters:
        - name: id
          in: path
          required: true
          description: ID del comentario
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentDto'
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para modificar este comentario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Comentario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      tags:
        - Comments
      summary: Eliminar comentario
      description: Eliminar un comentario (autor o Admin)
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          description: ID del comentario
          schema:
            type: integer
      responses:
        '204':
          description: Comentario eliminado exitosamente
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Sin permisos para eliminar este comentario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Comentario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ==================== CATALOGS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido mediante el endpoint `/api/auth/login`.
        
        Formato: `Bearer {token}`
        
        Ejemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    CommentDto:
      type: object
      properties:
        id:
          type: integer
          description: ID del comentario
          example: 1
        content:
          type: string
          description: Contenido del comentario
          example: "He revisado el error y parece ser un problema de configuración"
        createdById:
          type: integer
          description: ID del autor
          example: 2
        createdByName:
          type: string
          description: Nombre del autor
          example: "María García"
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación
          example: "2025-11-21T11:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Fecha de última edición
          example: "2025-11-21T11:00:00Z"
    CreateCommentDto:
      type: object
      required:
        - ticketId
        - content
      properties:
        ticketId:
          type: integer
          description: ID del ticket
          example: 1
        content:
          type: string
          maxLength: 2000
          description: Contenido del comentario
          example: "He revisado el error y encontré la causa"
    UpdateCommentDto:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 2000
          description: Nuevo contenido del comentario
          example: "Contenido actualizado del comentario"

    # ==================== CATALOG SCHEMAS ====================
...
    ProblemDetails:
      type: object
      description: Formato de error según RFC 7807
      properties:
        type:
          type: string
          description: URI que identifica el tipo de error
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          description: Título del error
          example: "Bad Request"
        status:
          type: integer
          description: Código HTTP
          example: 400
        detail:
          type: string
          description: Descripción detallada del error
          example: "El campo 'email' es requerido"
        instance:
          type: string
          description: URI de la instancia del error
          example: "/api/auth/login"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Errores de validación por campo
          example:
            email:
              - "El campo 'email' es requerido"
              - "El formato del email es inválido"
            password:
              - "La contraseña debe tener al menos 6 caracteres"
        traceId:
          type: string
          description: ID de traza para debugging
          example: "00-abc123def456-789xyz-00"